Do not show time zone notifications if the time zone is first set
This CL prevents showing a timezone notification when the timezone is first set, e.g. after setup wizard.

It does this by adding the old confidence to TimeZoneChangeEvent and only triggering the notification if the old confidence is not zero. This
prevents showing a notification if there was no previous confidence.

Test: flash + wipe device, no notification
Test: atest NotifyingTimeZoneChangeListenerTest.java
Flag: EXEMPT bugfix
Bug: b/421857844
Change-Id: Ief0425aa9aaef708eaff58eb9dfa7d83c6acf371