Added check for SDP and non-PD power ports.
For SDP BC1.2, this requires data line for initial handshake. To accommodate this, additional checks are added to see if a connected USB port for non-PD or BC1.2 compliance (which happens when USB data lines are disabled) and select the appropriate notification.

Test: manual
Bug: 420667983
Flag: android.security.aapm_feature_usb_data_protection
Change-Id: Ib6369cf4e58ac0e498e4cbd6007895db4fcd80fe