Resume app switching when app process is restarted
Process restart may require trampoline activity launch, for which
app switching needs to be enabled. App switching may be allowed
only for specific cases while/after Recents is shown.

As when a process is restarted, the user should be interacting
with the app, so it's safe to do this.

Flag: EXEMPT bug fix
Bug: 421048151
Test: CompatModeTests
Change-Id: Iff2accc543549a5a22a78705f886848530443f72