Adjust the IAMF AAC/Opus bitrates to be dependent on the number of channels.
Previously, we hardcoded the bitrate range based on the bitrate ranges for the separate AAC and Opus decoders, but IAMF can have many more channels.  So we scale the bitrate range with number of channels, set by profile, but essentially, we are only rejecting very large numbers.

Bug: 411204259
Flag: EXEMPT bug fix
Change-Id: I61ac12e1261abf77f01db89d6550bb3d9458354e