[PiP2] Update insets when rotating display layout
PipDisplayLayoutState#rotateTo() is a method we use
in PiP transitions (primarily enter with fixed rotation)
to calculate the display bounds at some target rotation
(in async rotation scheme the actual display rotation hasn't
happened yet).

However, PipDisplayLayoutState also caches the insets due
to the navbar to calculate the right insetsBounds, which is
ultimately what is used to get PiP entry bounds.

Therefore, whenever we rotate the display layout in-place,
we also need to make sure we rotate the insets to avoid extra
offsets to PiP on the wrong sides.

Bug: 418250885
Flag: com.android.wm.shell.enable_pip2
Test: manually enter PiP from ROTATION_270 and _90 in btn nav
Change-Id: I1fbcc6a0847f0ddfdab154d32d07ffc342b400dc