Skip bubbled tasks from merging into recents transition
We skip tasks with alwaysOnTop set to true from merging into the recents
transition.
When bubbles are collapsing, we remove the alwaysOnTop flag before the
transition starts playing.
When going from bubbles to recents, we can have the recents transition
start playing and then the bubbles collapse transition starts. In this
case we would attempt to merge the bubbles collapse transition into the
recents transition. But the bubbles transition does not support this.
Instead of checking for alwaysOnTop flag, add a check if the task is a
bubble. If it is, skip it from merging into the recents transition.

Bug: 418760100
Test: atest WMShellUnitTests:RecentsTransitionHandlerTest
Flag: com.android.wm.shell.fix_bubbles_to_recents
Change-Id: I1504f6d78397925bad02495942e54efe499c7f0f