Fallback to home activity from packageManager if supplier returns null
On certain devices home activity is never updated from system ui so the
supplier always returns null. In this case default to getting the home
activity from the package manager to just return sys ui as the home
activity. This returns the behaviour to that of before the refatoring
of `DesktopModeCompatPolicy`

Flag: EXEMPT bug fix
Fixes: 424861043
Test: manual
Change-Id: I4e42ce5b684d95ccd7f30c57a610f223f96e798f