Add previously released APIs to android-non-updatable-doc-stubs*
Previously, the `android-non-updatable-doc-stubs*` `droidstubs` modules
did not provide a previously released API which meant that any flagged
APIs that were reverted to their previous state would be removed
because there was no previous state. That lead to problems when
changing existing APIs, e.g. to add annotations.

This change updates those module to provide the previously released
API. It uses `check_api.last_released...` as that is currently the only
way to provide a previously released API but it disabled the
compatibility checks as they are done elsewhere.

Some special handling of `system-server` was needed as the module
generates an `annotations.zip` file that includes `module-lib` APIs
too but `system-server` only extends `public` so its previously
released API does not include `module-lib`.

Flag: EXEMPT build fix to improve flag support
Bug: 425309457
Test: TH
Change-Id: Id6c20242d97ba3d1f783ac0cd16982be0a798778