Add semantics actions to BundleHeader.
This adds the handling for dismiss, expand, and collapse actions at the level of the BundleHeader node, which is the one that gets focus in the accessibility tree. It was previously inheriting its actions from the ExpandableNotificationRow, which meant that the actions were available but not being announced.

Fixes: 424345757
Bug: 425559441
Test: manual by selecting the bundle header using talkback and confirming available actions and dismiss/expand/collapse behavior
Flag: com.android.systemui.notification_bundle_ui
Change-Id: I7128960ce4199670b8be3f08ed286485e9196894