Adjust a11y announcement order for expanded notifs
This ensures that the app name and top line content is read before the
actual content of the notification.

This was not needed for base and progress, as they have slightly
different hierarchies. It also doesn't quite work for the old media
style, but hopefully there shouldn't really be any of those
notifications in the wild.

Fix: 425757452
Test: manual
Flag: EXEMPT bugfix
Change-Id: Ia7d9f6c5e315816176ce8e32c650dcc61a20cbbc