Disable device rotation on glanceable hub on mobile
Currently we explicitly make the lockscreen orientation aware when
glanceable hub is showing, which was to enable landscape mode on mobile.
This change removes this logic to have the glanceable hub follow the
default lockscreen orientation awareness, which is true for large
screens and false for mobile.

Test: tested orientation disabled on mobile but keeps enabled on tablets
Bug: 419357983
Flag: com.android.systemui.glanceable_hub_v2
Change-Id: I1a4755f15f65ae022839179d1bafaa9a6f80d4d5