[ADI][34/N] enable a verification policy delegate
The policy delegate is an app that is specified by the system at build
time that can change the developer verification policy on behalf of the
developer verifier. It is a fail-safe mechanism provided for situations
where the verifier cannot function properly.

This CL updates the permission check around getting and setting developer
verification policy because these two methods are now accessible by the
policy delegate app which may not have the developer verification
permission.

The CL also provides an API for the verifier to query the package name
of the policy delegate app.

API-Coverage-Bug: 420966373

BUG: 360129657
FLAG: android.content.pm.verification_service
Test: atest android.content.pm.cts.PackageInfoTest
Test: atest CtsRootDeveloperVerificationPrivInstallerTarget36TestCases CtsRootDeveloperVerificationInstallerTarget36TestCases CtsRootDeveloperVerificationPrivInstallerTarget35TestCases CtsRootDeveloperVerificationInstallerTarget35TestCases CtsRootDeveloperVerificationVerifierTestCases

Change-Id: I58ef05a85abd9f081f99de35775f596e289e3652