[SB] Never hide status bar if device isn't provisioned.
While a user is still setting up their device, never let the status bar
be hidden. This fix is a workaround for a deeper issue where the shade
is reporting itself as expanded during setup wizard, which is
unexpected.

No unit tests added in this CL because that will make this CL much
harder to cherrypick to the release, will add tests in a followup.

Bug: 418020209
Flag: EXEMPT bugfix
Test: In setup wizard, tap the status bar repeatedly -> verify the
status bar never hides
Test: smoke test of status bar on homescreen, shade, and lockscreen
Test: atest HomeStatusBarViewModelImplTest

Change-Id: I8cf4ceb4355a2c29ab9a5aad2dae64120252b2e0