Fix clock handling of configuration change
The `Configuration` instance passed to onConfigurationChange was always the same (in most cases), so caching it in the `Clock` class didn't work.

With this we're just caching the field we care about (instead of the entire configuration, which would be cleaner but less efficient), and comparing them one by one.

Bug: 429007403
Flag: com.android.systemui.shade_window_goes_around
Test: change font scale and density with `adb shell settings put system font_scale 2` and check the clock font size changes.
Change-Id: Id6694bdca813a53c5d6a2e35eec29d95d3d3b7c7