[80/n] Use single surface as default strategy
It's easier to handle multiple surfaces only in case the top activity
is transparent.

The heuristic will be improved knowing if the rounded corners have been
already applied on the main activity or if some action will be required in
Shell (e.g. specific rounded corners implementation in Shell).

Flag: EXEMPT Refactoring
Bug: 377875146
Test: atest WMShellUnitTests:LetterboxControllerStrategyTest

Change-Id: I827c887913ff2dfe53efee4e3f960f3cd6169061