Skip duplicated request of killing task process
If mKillProcessesOnDestroyed is set, the process will be killed when
the last activity of the task is destroyed. So it is unnecessary to
evaluate the process to kill if removeTask is called again for the
same task. Otherwise because the top activity of the target task is
already finishing, killTaskProcessesIfPossible may be called again
and clear mKillProcessesOnDestroyed unexpectedly when waiting for
the activities to be destroyed, which could cancel the kill.

Bug: 431659784
Flag: EXEMPT bugfix
Test: Launch a package in a different task. Invoke removeTask twice
      from a different package. The process of the target task should
      be killed.
Change-Id: I2501ee99c2804ea823532a90c084eaeb29abc03a