[82/n] Improving Skip Logic
The actual skipping logic needs to be improved because there are cases
when the Changes cannot be completely skipped but they require
different actions depending on other factors. For instance, in some
cases, the Changes should be ignored when the letterbox surfaces are created
but not ignored in case of resizing (e.g. Split Screen).

Flag: EXEMPT Refactoring
Bug: 430486865
Test: atest WMShellUnitTests:LetterboxLifecycleEventTest
Test: atest WMShellUnitTests:SkipLetterboxLifecycleEventFactoryTest

Change-Id: I6cf0762c3034bc3251805e6089825177c05c834e