Hide status bar if keyguard is visible.
The legacy isHomeStatusBarAllowed flow doesn't cover edge cases like canceled swipe to dismiss transitions, and we don't have enough information in the KeyguardTransitionFramework to determine whether we're in a swipe/canceled swipe.

keyguard_wm_state_refactor and/or flexiglass do have this information, so this is a temporary fix.

Fixes: 407865650
Test: swipe to unlock partially
Test: cancel a swipe to unlock
Flag: EXEMPT bugfix
Change-Id: I2ea4c7ac9fc2042d55f305dc4f89322f53820e21